+++ website/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies [3533e00]
+  - @chakra-ui/vue@0.6.2
+  - @chakra-ui/theme-vue@0.2.7
+
+## 0.5.2-next.0
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @chakra-ui/vue@0.6.2-next.0
+
 ## 0.5.1
 
 ### Patch Changes